+++ tests/release/neg-missing-releaseplanadmission.go
@@ -1,9 +1,9 @@
 package release
 
 import (
+	"github.com/redhat-appstudio/application-api/api/v1alpha1"
 	"strings"
 
-	ecp "github.com/enterprise-contract/enterprise-contract-controller/api/v1alpha1"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
@@ -18,7 +18,6 @@ var _ = framework.ReleaseSuiteDescribe("[HACBS-2360] Release CR fails when missi
 	var err error
 
 	var devNamespace, managedNamespace string
-	var ecPolicy ecp.EnterpriseContractPolicySpec
 
 	var releaseCR *releaseApi.Release
 	AfterEach(framework.ReportFailure(&fw))
@@ -34,21 +33,9 @@ var _ = framework.ReleaseSuiteDescribe("[HACBS-2360] Release CR fails when missi
 		_, err = fw.AsKubeAdmin.CommonController.CreateTestNamespace(managedNamespace)
 		Expect(err).NotTo(HaveOccurred(), "Error when creating namespace '%s': %v", managedNamespace, err)
 
-		// Get the Spec from the default EnterpriseContractPolicy. This resource has up to date
-		// references and contains a small set of policy rules that should always pass during
-		// normal execution.
-		defaultECP, err := fw.AsKubeAdmin.TektonController.GetEnterpriseContractPolicy("default", "enterprise-contract-service")
-		Expect(err).NotTo(HaveOccurred(), "Error when fetching the default ECP: %v", err)
-		ecPolicy = defaultECP.Spec
-		_, err = fw.AsKubeAdmin.HasController.CreateApplication(applicationName, devNamespace)
-		Expect(err).NotTo(HaveOccurred())
-		_, err = fw.AsKubeAdmin.IntegrationController.CreateSnapshotWithComponents(snapshotName, "", applicationName, devNamespace, snapshotComponents)
-		Expect(err).NotTo(HaveOccurred())
-		_, err = fw.AsKubeAdmin.ReleaseController.CreateReleaseStrategy(releaseStrategyName, managedNamespace, releasePipelineName, releasePipelineBundle, releaseStrategyPolicy, serviceAccount, paramsReleaseStrategy)
+		_, err = fw.AsKubeAdmin.IntegrationController.CreateSnapshotWithComponents(snapshotName, "", applicationName, devNamespace, []v1alpha1.SnapshotComponent{})
 		Expect(err).NotTo(HaveOccurred())
 		_, err = fw.AsKubeAdmin.ReleaseController.CreateReleasePlan(sourceReleasePlanName, devNamespace, applicationName, managedNamespace, "")
-		Expect(err).NotTo(HaveOccurred())
-		_, err = fw.AsKubeAdmin.TektonController.CreateEnterpriseContractPolicy(releaseStrategyPolicy, managedNamespace, ecPolicy)
 		Expect(err).NotTo(HaveOccurred())
 		_, err = fw.AsKubeAdmin.ReleaseController.CreateRelease(releaseName, devNamespace, snapshotName, sourceReleasePlanName)
 		Expect(err).NotTo(HaveOccurred())
