+++ tests/release/pipelines/const.go
@@ -1,4 +1,4 @@
-package release
+package pipelines
 
 import (
 	"time"
@@ -8,30 +8,32 @@ import (
 )
 
 const (
-	snapshotName                                = "snapshot"
-	destinationReleasePlanAdmissionName         = "sre-production"
-	releaseName                                 = "release"
-	applicationName                             = "application"
 	serviceAccount                              = constants.DefaultPipelineServiceAccount
-	releaseStrategyPolicy                       = "policy"
-	verifyEnterpriseContractTaskName            = "verify-enterprise-contract"
 	applicationNameDefault               string = "appstudio"
 	componentName                        string = "dc-metro-map"
 	releaseStrategyPolicyDefault         string = "mvp-policy"
-	releasePipelineServiceAccountDefault string = "release-service-account"
+	releaseStrategyServiceAccountDefault string = "release-service-account"
 	sourceReleasePlanName                string = "source-releaseplan"
 	targetReleasePlanAdmissionName       string = "demo"
 	releasePvcName                       string = "release-pvc"
+	releaseEnvironment                   string = "production"
 	redhatAppstudioUserSecret            string = "hacbs-release-tests-token"
 	hacbsReleaseTestsTokenSecret         string = "redhat-appstudio-registry-pull-secret"
 	publicSecretNameAuth                 string = "cosign-public-key"
 	gitSourceComponentUrl                string = "https://github.com/scoheb/dc-metro-map"
 	releasedImagePushRepo                string = "quay.io/redhat-appstudio-qe/dcmetromap"
-	cacheSyncTimeout                            = 1 * time.Minute
-	releaseCreationTimeout                      = 5 * time.Minute
-	releasePipelineRunCreationTimeout           = 10 * time.Minute
-	releasePipelineRunCompletionTimeout         = 20 * time.Minute
-	releasePlanOwnerReferencesTimeout           = 1 * time.Minute
+	additionalReleasedImagePushRepo      string = "quay.io/redhat-appstudio-qe/simplepython"
+
+	additionalComponentName             string = "simple-python"
+	additionalGitSourceComponentUrl     string = "https://github.com/devfile-samples/devfile-sample-python-basic"
+	pyxisStageImagesApiEndpoint         string = "https://pyxis.preprod.api.redhat.com/v1/images/id/"
+	snapshotCreationTimeout                    = 5 * time.Minute
+	releaseCreationTimeout                     = 5 * time.Minute
+	releasePipelineRunCreationTimeout          = 10 * time.Minute
+	releasePipelineRunCompletionTimeout        = 20 * time.Minute
+	avgControllerQueryTimeout                  = 5 * time.Minute
+	releaseDeploymentTimeout                   = 10 * time.Minute
+	releaseFinishedTimeout                     = 5 * time.Minute
 
 	defaultInterval = 100 * time.Millisecond
 )
